+++ src/components/basic_table/basic_table.tsx
@@ -477,7 +477,7 @@ export class EuiBasicTable<T = any> extends Component<
     );
 
     const table = this.renderTable();
-    const paginationBar = this.renderPaginationBar(items.length);
+    const paginationBar = this.renderPaginationBar();
 
     return (
       <div className={classes} {...rest}>
@@ -1235,9 +1235,9 @@ export class EuiBasicTable<T = any> extends Component<
     return profile.align;
   }
 
-  renderPaginationBar(itemsLength: number) {
+  renderPaginationBar() {
     const { error, pagination, onChange } = this.props;
-    if (!error && pagination && itemsLength > 0) {
+    if (!error && pagination && pagination.totalItemCount > 0) {
       if (!onChange) {
         throw new Error(`The Basic Table is configured with pagination but [onChange] is
         not configured. This callback must be implemented to handle pagination changes`);