+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 **Bug Fixes**
 
 - Fixed `EuiSuperSelect` not rendering full width when `isOpen` is `true` ([#3495](https://github.com/elastic/eui/pull/3495))
+- Fixed `EuiBasicTable` shows no items if all items of last page is deleted  ([#3422](https://github.com/elastic/eui/pull/3422))
 
 **Deprecations**